fix: 修正艺术家、收藏和主页的播放次数计算逻辑

+++ reverie-ui/src/pages/artists.rs
@@ -20,7 +20,7 @@ pub fn ArtistsPage() -> Element {
             .map(|i| Artist {
                 id: format!("artist-{}", i),
                 name: format!("Artist {}", i),
-                album_count: 3 + (i % 5) as i32,
+                album_count: 3 + (i % 5),
                 cover_art: None,
                 artist_image_url: None,
                 starred: None,
